chore(suite): split Password component into multiple components
