+++ packages/suite/src/views/settings/SettingsDebug/PasswordManager/PasswordManager.tsx
@@ -8,10 +8,11 @@ import { selectDevice } from '@suite-common/wallet-core';
 import { SectionItem, TextColumn, ActionColumn } from 'src/components/suite';
 import { useSelector, useDispatch } from 'src/hooks/suite';
 import * as metadataActions from 'src/actions/suite/metadataActions';
-import * as metadataUtils from 'src/utils/suite/metadata';
 import type { PasswordEntry } from 'src/types/suite/metadata';
 import { METADATA } from 'src/actions/suite/constants';
 import { selectSelectedProviderForPasswords } from 'src/reducers/suite/metadataReducer';
+
+import { PasswordEntry as PasswordEntryComponent } from './PasswordEntry';
 
 const PasswordsList = styled.div`
     display: flex;
@@ -20,101 +21,12 @@ const PasswordsList = styled.div`
     flex: 1;
 `;
 
-const PasswordEntryRow = styled.div`
-    margin-bottom: 4px;
-`;
-
-const PasswordEntryBody = styled.div`
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    margin-left: 8px;
-`;
-
-const PasswordEntryTitle = styled.div`
-    font-size: 14px;
-`;
-
-const PasswordEntryUsername = styled.div`
-    font-size: 14px;
-`;
-
-const PasswordEntryPassword = styled.div`
-    font-size: 14px;
-`;
-
 const FILENAME_MESS = '5f91add3fa1c3c76e90c90a3bd0999e2bd7833d06a483fe884ee60397aca277a';
 const HD_HARDENED = 0x80000000;
 const PATH = [10016 + HD_HARDENED, 0];
 const DEFAULT_KEYPHRASE = 'Activate TREZOR Password Manager?';
 const DEFAULT_NONCE =
     '2d650551248d792eabf628f451200d7f51cb63e46aadcbb1038aacb05e8c8aee2d650551248d792eabf628f451200d7f51cb63e46aadcbb1038aacb05e8c8aee';
-
-interface PasswordEntryProps extends PasswordEntry {
-    devicePath: string;
-}
-
-const getDisplayKey = (title: string, username: string) =>
-    // todo: implement this for the other category too: https://github.com/trezor/trezor-password-manager/blob/6266f685226bc5d5e0d8c7f08490b282f64ad1d1/source/background/classes/trezor_mgmt.js#L389-L390
-    `Unlock ${title} for user ${username}?`;
-
-const PasswordEntryComponent = ({
-    username,
-    title,
-    nonce,
-    password,
-    devicePath,
-}: PasswordEntryProps) => {
-    const [decodedPassword, setDecodedPassword] = useState('');
-    const [inProgress, setInProgress] = useState(false);
-
-    const decode = useCallback(() => {
-        if (inProgress) return;
-        setInProgress(true);
-        TrezorConnect.cipherKeyValue({
-            device: { path: devicePath },
-            override: true,
-            useEmptyPassphrase: true,
-            path: PATH,
-            key: getDisplayKey(title, username),
-            value: nonce,
-            encrypt: false,
-            askOnEncrypt: false,
-            askOnDecrypt: true,
-        })
-            .then(result => {
-                if (result.success) {
-                    const decrypted = metadataUtils.decrypt(
-                        Buffer.from(password.data),
-                        Buffer.from(result.payload.value, 'hex'),
-                    );
-                    setDecodedPassword(decrypted);
-                }
-            })
-            .finally(() => {
-                setInProgress(false);
-            });
-    }, [username, title, inProgress, devicePath, nonce, password]);
-
-    return (
-        <>
-            <PasswordEntryRow>
-                <PasswordEntryBody>
-                    <PasswordEntryTitle>{title}</PasswordEntryTitle>
-                    <PasswordEntryUsername>{username}</PasswordEntryUsername>
-                    <PasswordEntryPassword>
-                        {!decodedPassword ? (
-                            <Button onClick={decode} type="button" variant="tertiary">
-                                {inProgress ? '....' : 'decode'}
-                            </Button>
-                        ) : (
-                            decodedPassword
-                        )}
-                    </PasswordEntryPassword>
-                </PasswordEntryBody>
-            </PasswordEntryRow>
-        </>
-    );
-};
 
 export const PasswordManager = () => {
     // todo: filename is saved only locally. for production grade state of this feature we will of course need to save it into app state.
